# Simple serializer