--- a/event/serialize.go
+++ b/event/serialize.go
@@ -1,16 +1,14 @@
 package event
 
-import (
-	"github.com/mesg-foundation/engine/hash/serializer"
-)
+import "github.com/mesg-foundation/engine/hash/hashserializer"
 
-func (data *Event) Serialize() string {
+func (data *Event) HashSerialize() string {
 	if data == nil {
 		return ""
 	}
-	ser := serializer.New()
+	ser := hashserializer.New()
 	ser.AddString("2", data.InstanceHash.String())
 	ser.AddString("3", data.Key)
 	ser.Add("4", data.Data)
-	return ser.Serialize()
+	return ser.HashSerialize()
 }