--- a/event/serialize.go
+++ b/event/serialize.go
@@ -7,9 +7,9 @@ func (data *Event) HashSerialize() string {
 	if data == nil {
 		return ""
 	}
-	ser := hashserializer.New()
-	ser.AddString("2", data.InstanceHash.String())
-	ser.AddString("3", data.Key)
-	ser.Add("4", data.Data)
-	return ser.HashSerialize()
+	return hashserializer.New().
+		AddString("2", data.InstanceHash.String()).
+		AddString("3", data.Key).
+		Add("4", data.Data).
+		HashSerialize()
 }