--- a/hash/hashserializer/hashserializer.go
+++ b/hash/hashserializer/hashserializer.go
@@ -1,9 +1,98 @@
 // Package hashserializer creates a string representing arbitrary value that can then be used to generate a deterministic hash.
 //
-// Each value can be added to the HashSerializer using one of the AddXXX function alongside a prefix.
-// The prefix should be unique and the AddXXX function should always be called in the same order in order to create a deterministic string.
+// Each value must have a unique prefix.
 //
-// Empty value are discarded and its prefix will not be part of the generated string.
+// Empty or default value are discarded.
+//
+// The order to add the value should always be the same if you want the created string to be deterministic.
+//
+// Each value can be added to the HashSerializer using one of the AddXXX function alongside the prefix:
+//  ser := New()
+//  ser.AddBool(prefix, value)
+//  ser.AddFloat(prefix, value)
+//  ser.AddInt(prefix, value)
+//  ser.AddString(prefix, value)
+//  ser.AddStringSlice(prefix, value)
+//  ser.Add(prefix, value) // value must implement function `HashSerialize() string`
+//  hashSerialized := ser.HashSerialize()
+//
+// Format
+//
+// The format of the generated string for one value with a prefix is:
+//  prefix:value;
+// where value is the string representation of the value.
+//
+// When adding multiple values, they are append together:
+//  prefix1:value1;prefix2:value2;
+//
+// Empty and default value
+//
+// Empty and default values are discarded and not part of the generated string. For slices and maps, if every value is empty then the slice or maps is also considered as empty.
+//
+// The following values are all considered as empty and will not be part of the generated string:
+//  string("")
+//  bool(false)
+//  int(0)
+//  float(0.0)
+//  []string{}
+//  []string{""}
+//  map[string]string{}
+//  map[string]string{"a":"", "b":""}
+//
+// String
+//
+// String values are the simplest are the value doesn't need to be encoded.
+//
+// For prefix "foo" and value "bar", the result is:
+//  foo:bar;
+//
+// Bool
+//
+// Only bool value true produce an output as the default value false is discarded.
+//
+// For prefix "foo" and value "true", the result is:
+//  foo:true;
+//
+// Int
+//
+// Int values are encoded to string using base 10 representation. If the value is negative, the sign "-" should prefix the value.
+//
+// For prefix "foo" and value "42", the result is:
+//  foo:42;
+//
+// Float
+//
+// Float values are encoded to string using base 10 with no exponent. If the value is negative, the sign "-" should prefix the value.
+//
+// For prefix "foo" and value "3.14159265359", the result is:
+//  foo:3.14159265359;
+//
+// Slice
+//
+// For slices, prefixes are the index of the elements. Empty value are discarded.
+//
+//  []string{"foo", "", "bar"}
+//  0:foo;2:bar;
+//
+// Map
+//
+// For maps, prefixes are the keys. The maps must be sorted by keys to get a deterministic result. Empty value are discarded.
+//
+//  map[string]string{"c":"bar", "b": "", "a":"foo"}
+//  a:foo;c:bar;
+//
+// Struct
+//
+// Struct should implement the function `HashSerialize() string` to return the hash-serialized string.
+//
+// Nested
+//
+// When adding a struct, a slice or a map, the value is hash-serialized before added to the string.
+//
+// For prefix "foo" and value `[]string{"foo", "", "bar"}`, the result is:
+//  foo:0:foo;2:bar;;
+//
+//
 package hashserializer
 
 import (