--- a/instance/serialize.go
+++ b/instance/serialize.go
@@ -1,15 +1,13 @@
 package instance
 
-import (
-	"github.com/mesg-foundation/engine/hash/serializer"
-)
+import "github.com/mesg-foundation/engine/hash/hashserializer"
 
-func (data *Instance) Serialize() string {
+func (data *Instance) HashSerialize() string {
 	if data == nil {
 		return ""
 	}
-	ser := serializer.New()
+	ser := hashserializer.New()
 	ser.AddString("2", data.ServiceHash.String())
 	ser.AddString("3", data.EnvHash.String())
-	return ser.Serialize()
+	return ser.HashSerialize()
 }