--- a/instance/serialize.go
+++ b/instance/serialize.go
@@ -7,8 +7,8 @@ func (data *Instance) HashSerialize() string {
 	if data == nil {
 		return ""
 	}
-	ser := hashserializer.New()
-	ser.AddString("2", data.ServiceHash.String())
-	ser.AddString("3", data.EnvHash.String())
-	return ser.HashSerialize()
+	return hashserializer.New().
+		AddString("2", data.ServiceHash.String()).
+		AddString("3", data.EnvHash.String()).
+		HashSerialize()
 }