--- a/ownership/serialize.go
+++ b/ownership/serialize.go
@@ -1,15 +1,13 @@
 package ownership
 
-import (
-	"github.com/mesg-foundation/engine/hash/serializer"
-)
+import "github.com/mesg-foundation/engine/hash/hashserializer"
 
-func (data *Ownership) Serialize() string {
+func (data *Ownership) HashSerialize() string {
 	if data == nil {
 		return ""
 	}
-	ser := serializer.New()
+	ser := hashserializer.New()
 	ser.AddString("2", data.Owner)
 	ser.AddString("3", data.ResourceHash.String())
-	return ser.Serialize()
+	return ser.HashSerialize()
 }