… e2e/instance_test.go
… event/serialize_test.go
… execution/serialize_test.go
… hash/hash_test.go
… hash/hashserializer/hashserializer_test.go
… hash/serializer/serializer_test.go
… hash/structhash/structhash.go
… hash/structhash/structhash_test.go
… instance/serialize_test.go
… ownership/serialize_test.go
… process/serialize.go
… process/serialize_test.go
… protobuf/types/serialize.go
… protobuf/types/serialize_test.go
… protobuf/types/service.proto
… protobuf/types/struct_test.go
… runner/builder/builder.go
… runner/builder/container.go
… runner/serialize.go
… runner/serialize_test.go
… service/serialize.go
… service/serialize_test.go
… service/service.pb.go