Simplification of https://github.com/mesg-foundation/engine/pull/1678
Closes https://github.com/mesg-foundation/engine/issues/1667

This PR implements a simple and very performant serializer that requires every struct to implement a `func Serialize() string`.

I tried different implementations and compared them in terms of performance and UX of its API.
- hashing child-structures is a huge performance loss
- sorting is a medium performance loss
- using multiple times `buffer.WriteString` instead of constructing string with `fmt.Sprintf` is a medium performance improvement

I also implement the same serializer in JS to make sure it's also simple to implement and check the results.

I ran some benchmarks on a simple service struct, the difference from the less optimize from the most optimize is a x10 of performance improvement.

Here are the benchmarks i ran:
```
// with base58
// BenchmarkSerialize-4   	   45284	     26965 ns/op	    7928 B/op	     216 allocs/op
// BenchmarkSerialize-4   	   44287	     27114 ns/op	    7928 B/op	     216 allocs/op
// BenchmarkSerialize-4   	   45910	     26025 ns/op	    7928 B/op	     216 allocs/op

// without base58
// BenchmarkSerialize-4   	   98306	     11453 ns/op	    5848 B/op	     188 allocs/op
// BenchmarkSerialize-4   	   95521	     11667 ns/op	    5848 B/op	     188 allocs/op
// BenchmarkSerialize-4   	  106831	     11040 ns/op	    5848 B/op	     188 allocs/op

// same as previous one + eliminate empty value as soon as possible
// BenchmarkSerialize-4   	  130498	      8952 ns/op	    4352 B/op	     149 allocs/op
// BenchmarkSerialize-4   	  121597	     10585 ns/op	    4352 B/op	     149 allocs/op
// BenchmarkSerialize-4   	  127125	      9357 ns/op	    4352 B/op	     149 allocs/op

// same as previous one + no sort
// BenchmarkSerialize-4   	  136657	      8135 ns/op	    4000 B/op	     138 allocs/op
// BenchmarkSerialize-4   	  133406	      8202 ns/op	    4000 B/op	     138 allocs/op
// BenchmarkSerialize-4   	  137749	      8333 ns/op	    4000 B/op	     138 allocs/op

// same as previous one + bytes
// BenchmarkSerialize-4   	  234852	      4824 ns/op	    3432 B/op	      86 allocs/op
// BenchmarkSerialize-4   	  224851	      4972 ns/op	    3432 B/op	      86 allocs/op
// BenchmarkSerialize-4   	  219336	      5380 ns/op	    3432 B/op	      86 allocs/op

// same as previous one + buffer pool
// BenchmarkSerialize-4   	  244681	      4571 ns/op	    2328 B/op	      73 allocs/op
// BenchmarkSerialize-4   	  246154	      4622 ns/op	    2328 B/op	      73 allocs/op
// BenchmarkSerialize-4   	  237271	      4530 ns/op	    2328 B/op	      73 allocs/op

// same as previous one + return early on empty pairs
// BenchmarkSerialize-4   	  263187	      4401 ns/op	    2328 B/op	      73 allocs/op
// BenchmarkSerialize-4   	  236986	      4268 ns/op	    2328 B/op	      73 allocs/op
// BenchmarkSerialize-4   	  279604	      4284 ns/op	    2328 B/op	      73 allocs/op

// same as previous one + no sort
// BenchmarkSerialize-4   	  291133	      3688 ns/op	    1976 B/op	      62 allocs/op

// same as previous one + remove intermediate array to hold the key/value, direct write to a dedicated buffer (no buffer pool)
// BenchmarkSerialize-4   	  468823	      2347 ns/op	    2000 B/op	      30 allocs/op
// BenchmarkSerialize-4   	  459656	      2329 ns/op	    2000 B/op	      30 allocs/op
// BenchmarkSerialize-4   	  458227	      2260 ns/op	    2000 B/op	      30 allocs/op
```


---

TODO:
- [ ] implement serializer test in every structure