@antho1404 2020-02-29
Didn't review in detail but really like the direction, for sure, more code duplication but a lot more control and it is now really simple to understand. 

(Blocking):
One thing that I really don't like, this is not really a generic serialization method so having `Serialize` as function name is quite bad. I would be ok with `hashserialize` or something like that. 

(Not blocking but definitely really nice to have):
Also I hear the performance issue with hashing all the sub structs but except in the process that we might have 10+ sub structs, this is negligible and I would definitely go with a function `Hash` that does the serialization internally and returns the hash this would be really easy to understand and use.