@NicolasMahe 2020-03-04
> (Blocking):
> One thing that I really don't like, this is not really a generic serialization method so having `Serialize` as function name is quite bad. I would be ok with `hashserialize` or something like that.

done 👌 

> (Not blocking but definitely really nice to have):
> Also I hear the performance issue with hashing all the sub structs but except in the process that we might have 10+ sub structs, this is negligible and I would definitely go with a function `Hash` that does the serialization internally and returns the hash this would be really easy to understand and use.

I would like to not use hash inside the `hashserialize` function because:
- it is not needed and it adds one more step
- it creates a different behavior for "simple" value and "complex" value (why not hashing every value?)
- I don't want to force a specific string representation of the hash (we are already discussing about using bech32 instead of base58 https://github.com/mesg-foundation/engine/issues/1700)
- It creates a dependency cycle if `hashserializer` is using `hash` package