@NicolasMahe 2020-03-04
I updated the unit test to example test to normally generate a really nice documentation.
But because it's not released, I cannot check. @krhubert do you know how to display something like https://pkg.go.dev/mod/github.com/mesg-foundation/engine but locally?