@krhubert 2020-03-04
@NicolasMahe I don't know if go.dev can be run locally, because it's in beta, but for sure you can run someting like godoc.org with https://godoc.org/golang.org/x/tools/cmd/godoc