@antho1404 2020-03-04
modules are not supported with `godoc -http=":6060"` unless you are running go v1.14