@NicolasMahe 2020-03-05
> modules are not supported with `godoc -http=":6060"` unless you are running go v1.14.
> There is a workaround for it otherwise https://stackoverflow.com/questions/54933983/how-to-serve-documentation-using-godoc-together-with-go-modules

I install go v1.14 and I was not able to access the engine's package using `godoc -http=":6060"`.
The workaround you link works! I created a `doc.sh` script to run it.