@NicolasMahe 2020-03-05
Documentation improved!
To check it, run in the engine folder:
```
godoc -http=:6060
```
and open: http://localhost:6060/pkg/github.com/mesg-foundation/engine/hash/hashserializer