@NicolasMahe 2020-03-10
POC js implementation: https://github.com/mesg-foundation/js-sdk/pull/182